Reset cookies each WebView appears
Instagram may sometimes ask users to confirm their account with a code from their email and if there is a persistent cookie in the WebView browser, then instagram may not load the login page for some reason, so i figured out the best way to deal this so far is to clear out the cookies before showing the Instagram login webview.

Although there is a bug where a user is required to press "share" again, while I set up the logic in a way that re-extracting shared items should happen automatically after a user dismissed the Instagram login WebView sheet with a correct credentials.